DRTII-852 Refactor for tidying and removal of duplicate code
