+++ server/src/main/scala/services/exports/flights/templates/FlightsExport.scala
@@ -33,6 +33,10 @@ trait FlightsExport {
 
   val request: FlightsRequest
 
+  val standardFilter: (Terminal, ApiFlightWithSplits) => Boolean = (terminal, fws) => fws.apiFlight.Terminal == terminal
+
+  val flightsFilter: (Terminal, ApiFlightWithSplits) => Boolean
+
   val queueNames: Seq[Queue] = ApiSplitsToSplitRatio.queuesFromPaxTypeAndQueue(PaxTypesAndQueues.inOrder)
 
   def flightToCsvRow(fws: ApiFlightWithSplits): String = rowValues(fws).mkString(",")
@@ -68,6 +72,7 @@ trait FlightsExport {
     flightsStream.mapConcat { flights =>
       uniqueArrivalsWithCodeShares(flights.flights.values.toSeq)
         .map(_._1)
+        .filter(fws => flightsFilter(terminal, fws))
         .sortBy { fws =>
           val arrival = fws.apiFlight
           (arrival.PcpTime, arrival.VoyageNumber.numeric, arrival.Origin.iata)
