+++ server/src/main/scala/services/exports/flights/templates/FlightsWithSplitsWithActualApiExport.scala
@@ -1,9 +1,6 @@
 package services.exports.flights.templates
 
 import actors.PartitionedPortStateActor.{FlightsRequest, GetFlightsForTerminalDateRange}
-import akka.NotUsed
-import akka.stream.scaladsl.Source
-import drt.shared.FlightsApi.FlightsWithSplits
 import drt.shared.Queues.Queue
 import drt.shared.Terminals._
 import drt.shared._
@@ -20,17 +17,6 @@ trait FlightsWithSplitsWithActualApiExport extends FlightsWithSplitsExport {
     actualAPISplitsForFlightInHeadingOrder(fws, actualApiHeadings).toList).map(s => s"$s")
 }
 
-case class FlightsWithSplitsWithActualApiExportImpl(start: SDateLike, end: SDateLike, terminal: Terminal) extends FlightsWithSplitsWithActualApiExport
-
-case class LHRFlightsWithSplitsWithActualApiExportImpl(start: SDateLike, end: SDateLike, terminal: Terminal) extends FlightsWithSplitsWithActualApiExport with LHRFlightsWithSplitsExportWithDiversions {
-  override def filterAndSort(flightsStream: Source[FlightsWithSplits, NotUsed]): Source[ApiFlightWithSplits, NotUsed] =
-    flightsStream.mapConcat { flights =>
-      uniqueArrivalsWithCodeShares(flights.flights.values.toSeq)
-        .map(_._1)
-        .filter(fws => filter(terminal, fws))
-        .sortBy { fws =>
-          val arrival = fws.apiFlight
-          (arrival.PcpTime, arrival.VoyageNumber.numeric, arrival.Origin.iata)
-        }
-    }
+case class FlightsWithSplitsWithActualApiExportImpl(start: SDateLike, end: SDateLike, terminal: Terminal) extends FlightsWithSplitsWithActualApiExport {
+  override val flightsFilter: (Terminal, ApiFlightWithSplits) => Boolean = standardFilter
 }
